+++ cosmetics-web/app/models/concerns/privilies.rb
@@ -1,0 +1,1 @@
+module Priviliges; end
