+++ cosmetics-web/app/views/responsible_persons/notifications/_product_details.html.erb
@@ -30,7 +30,7 @@
       )
     ) %>
 
-    <% if current_user.can_view_product_ingredients? %>
+    <% if can_view_product_ingredients? %>
       <h3 class="govuk-heading-m govuk-!-margin-top-8 govuk-!-margin-bottom-2">
         Ingredients
       </h3>
